Place board in separate folder and extracted classes into separate file
