+++ src/app/backend/game.ts
@@ -1,7 +1,9 @@
-import {Board, Coordinate, RectangularArea} from './board';
+import {Board} from './board/board';
 import {BoardSerialization} from './board-serialization';
 import {GameState, GameStepAlgorithm} from './game-step-algorithm';
 import {Cell} from './cells';
+import {Coordinate} from './board/coordinate';
+import {RectangularArea} from './board/rectangular-area';
 
 /**
  * The entry point of using the game. Needs to be initialized using `readBoardFromString()`.
