+++ src/app/pages/canvas-prototype/canvas-drawer-facade.ts
@@ -1,7 +1,8 @@
 import {fabric} from 'fabric';
-import {Coordinate, RectangularArea} from '../../backend/board';
 import {Cell, CellWithDirection, Direction} from '../../backend/cells';
 import {CanvasCellImageCreator} from './canvas-cell-image-creator';
+import {Coordinate} from '../../backend/board/coordinate';
+import {RectangularArea} from '../../backend/board/rectangular-area';
 
 /**
  * Class that handles all the drawing of the canvas. Methods in this class are game-specific.
